Clean up global state after tests that modify it:

- reset sessions/environment in alchemy_db_utils_unittest.py
- reset environment in java_unittest.py and tools_dbmaint_unittest.py
- reset class member in job_unittest.py

+++ db_tests/job_unittest.py
@@ -226,9 +226,13 @@ CONFIG_FILE = "config.gem"
 
 class JobTestCase(unittest.TestCase):
 
+    def setUp(self):
+        self.job = None
+
     def tearDown(self):
         try:
-            os.remove(self.job.super_config_path)
+            if self.job:
+                os.remove(self.job.super_config_path)
         except OSError:
             pass
